refactor: rename `rulesEngine.types.js` to `ruleEngine.types.js`
