+++ src/core_modules/capture-core/rules/RuleEngine/types/ruleEngine.types.js
